requirements.txt
aiohttp==3.8.3
face-recognition==1.3.0
Pillow==9.2.0
requests==2.28.1
feedparser==6.0.10
pydantic~=1.10.2